+++ src/dct.lua
@@ -1,6 +1,5 @@
 -- SPDX-License-Identifier: LGPL-3.0
 
-local _G   = _G
 local dct = {
     _VERSION = "0.6",
     _DESCRIPTION = "DCT: DCS Dynamic Campaign Tools",
@@ -8,7 +7,7 @@ local dct = {
 }
 
 _G.dct = dct
-dct.settings  = require("dct.settings")(dctsettings)
+dct.settings  = require("dct.settings")()
 dct.init      = require("dct.init")
 dct.Theater   = require("dct.Theater")
